review: address PR #29202 feedback
- FileResponseStream: add optional on_abort callback (fires instead of
  on_complete on client disconnect). RequestContext routes it through
  the real onAbort so flags.aborted / additional_on_abort / req.signal
  fire as before.
- RequestContext.doSendfile: latch has_sendfile_ctx only after
  open/stat/dir validation succeeds, so an error handler returning a
  file-backed Response can re-enter doSendfile.
- FileRoute: ignore Range when route status_code != 200 (was serving
  partial body under custom status with Content-Range headers).
- RangeRequest: bytes=-N on a zero-length file returns .none (200/204)
  instead of .unsatisfiable (416), per RFC 9110 14.1.3.
- Move abort regression to test/regression/issue/20965.test.ts; drop
  the strict stderr assertion (ASAN JSC warnings).
