+++ test/js/bun/http/bun-serve-file.test.ts
@@ -1,6 +1,6 @@
 import type { Server } from "bun";
 import { afterAll, beforeAll, describe, expect, it, mock, test } from "bun:test";
-import { bunEnv, bunExe, normalizeBunSnapshot, rmScope, tempDir, tempDirWithFiles } from "harness";
+import { rmScope, tempDirWithFiles } from "harness";
 import { unlinkSync } from "node:fs";
 import { join } from "node:path";
 
@@ -639,58 +639,4 @@ describe.each([
     expect(res.status).toBe(200);
     expect(await res.text()).toBe(body);
   });
-});
-
-test("aborting a streaming file response mid-transfer does not leak pending_requests (server.stop resolves)", async () => {
-  using dir = tempDir("file-route-abort", {
-    "fixture.ts": /* ts */ `
-      import { connect } from "node:net";
-      import { writeFileSync } from "node:fs";
-      import { join } from "node:path";
-
-      const big = join(import.meta.dir, "big.bin");
-      writeFileSync(big, Buffer.alloc(5 * 1024 * 1024));
-
-      using server = Bun.serve({
-        port: 0,
-        routes: { "/big": new Response(Bun.file(big)) },
-        fetch: () => new Response("unreachable"),
-      });
-
-      // Raw TCP: send GET, receive first chunk, then drop the socket. The 5MB
-      // body fills the kernel send buffer so StreamTransfer is paused on
-      // backpressure when onAborted fires.
-      await new Promise<void>((resolve, reject) => {
-        const sock = connect({ port: server.port, host: "127.0.0.1" }, () => {
-          sock.write("GET /big HTTP/1.1\\r\\nHost: x\\r\\n\\r\\n");
-          sock.once("data", () => {
-            sock.destroy();
-            resolve();
-          });
-        });
-        sock.on("error", reject);
-      });
-
-      const result = await Promise.race([
-        server.stop().then(() => "stopped"),
-        Bun.sleep(2000).then(() => "HUNG: pending_requests was never decremented"),
-      ]);
-      console.log(result);
-      process.exit(result === "stopped" ? 0 : 1);
-    `,
-  });
-
-  await using proc = Bun.spawn({
-    cmd: [bunExe(), "fixture.ts"],
-    env: bunEnv,
-    cwd: String(dir),
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-
-  const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-
-  expect(normalizeBunSnapshot(stdout)).toBe("stopped");
-  expect(stderr).toBe("");
-  expect(exitCode).toBe(0);
 });
